+++ src/main/java/com/autotune/analyzer/recommendations/objects/MappedRecommendationForTerm.java
@@ -1,6 +1,5 @@
 package com.autotune.analyzer.recommendations.objects;
 
-import com.autotune.analyzer.recommendations.RecommendationConstants;
 import com.autotune.analyzer.recommendations.RecommendationNotification;
 
 import java.sql.Timestamp;
@@ -9,6 +8,6 @@ import java.util.HashMap;
 public interface MappedRecommendationForTerm {
     public HashMap<Integer, RecommendationNotification> getNotifications();
     public Timestamp getMonitoringStartTime();
-    public int getDurationInHrs();
+    public double getDurationInHrs();
     public MappedRecommendationForEngine getRecommendationByEngine(String EngineName);
 }
